Small performance update
Took the "hold on" out of the loop. 

The loop iterates on "plot" objects which are children of "axes", the method "hold" refers to the axes themselves so there's no need to call "hold on" in each iteration.

Improved the test case runtime of "niccases" from 8.998[s] to 5.238[s] on my machine.